Remove unnecessary files (controllers, views, routes). Do some refactoring of code.
* Views: Remove extra files
* Routes: Delete students, teachers and admins routes
* Model: Add validations to models
* Helpers: Use helper functions to simplify views

+++ app/controllers/users_controller.rb
@@ -3,10 +3,12 @@
 class UsersController < ApplicationController
   def index
     @users = User.all.order(:created_at)
+    authorize @users
   end
 
   def destroy
     @user = User.find(params[:id])
+    authorize @user, :index
 
     if @user.destroy
       redirect_to root_path, notice: 'User Removed'
